{"types":"C","runs":3}

+++ Problem19.java
@@ -2,7 +2,7 @@ import acm.program.ConsoleProgram;
 //19. კონსოლიდან წაიკითხეთ მთელი რიცხვი n და შემდეგ n ჯერ დაბეჭდეთ “hello ” + i, სადაც i  იცვლება 0 დან n-1 მდე.  
 public class Problem19 extends ConsoleProgram {
 	public void run(){
-		int n = readInt("Give me a number");
+		double n = readDouble("Give me a number");
 		for(int i=0; i<n; i++){
 			println("hello World"+i);
 		}
